pyproject.toml
[tool.poetry]
name = "mvp-clientes"
version = "0.1.0"
description = ""
authors = ["Diogo Lima"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
djangorestframework = "^3.13.1"
environs = {extras = ["django"], version = "^9.5.0"}
gunicorn = "^20.1.0"
django-cors-headers = "^3.13.0"
phonenumbers = "^8.12.54"
django-phonenumber-field = "^6.3.0"
Pillow = "^9.2.0"
twilio = "^7.13.0"
drf-spectacular = "^0.23.1"
psycopg2 = "^2.9.3"
platformdirs = "^2.5.2"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"
black = {version = "^22.6.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
